Wave System
Adds a wave counter, and makes random enemies spawn each wave. Enemies have a difficulty score, so harder enemies are worth 2 for example and easy enemies are worth 1. The wave will spawn enemies up until the difficulty score of all those enemies added together equals a difficulty cap, which increases the higher the wave number.
TLDR: waves get harder and have random enemies.

Everything is basic but can be added onto later when we know what enemies we want, and as long as everything is updated correctly it should work.